feat: CheckCertAndUIDAllowed_

+++ src/CraneCtld/RpcService/CtldGrpcServer.cpp
@@ -38,6 +38,8 @@ grpc::Status CraneCtldServiceImpl::SubmitBatchTask(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->task().uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   auto task = std::make_unique<TaskInCtld>();
   task->SetFieldsByTaskToCtld(request->task());
@@ -67,6 +69,8 @@ grpc::Status CraneCtldServiceImpl::SubmitBatchTasks(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->task().uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   std::vector<CraneExpected<std::future<task_id_t>>> results;
 
@@ -98,6 +102,8 @@ grpc::Status CraneCtldServiceImpl::CancelTask(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->operator_uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   *response = g_task_scheduler->CancelPendingOrRunningTask(*request);
   return grpc::Status::OK;
@@ -144,6 +150,8 @@ grpc::Status CraneCtldServiceImpl::QueryReservationInfo(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   auto res = g_account_manager->CheckUidIsAdmin(request->uid());
   if (!res) {
@@ -168,6 +176,8 @@ grpc::Status CraneCtldServiceImpl::ModifyTask(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   using ModifyTaskRequest = crane::grpc::ModifyTaskRequest;
 
@@ -263,6 +273,8 @@ grpc::Status CraneCtldServiceImpl::ModifyNode(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   auto res = g_account_manager->CheckUidIsAdmin(request->uid());
   if (!res) {
@@ -352,6 +364,8 @@ grpc::Status CraneCtldServiceImpl::ModifyPartitionAcl(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   CraneExpected<void> result;
 
@@ -439,6 +453,8 @@ grpc::Status CraneCtldServiceImpl::AddAccount(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   Account account;
   const crane::grpc::AccountInfo *account_info = &request->account();
@@ -471,6 +487,8 @@ grpc::Status CraneCtldServiceImpl::AddUser(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   User user;
   const crane::grpc::UserInfo *user_info = &request->user();
@@ -515,6 +533,9 @@ grpc::Status CraneCtldServiceImpl::AddQos(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
+
   Qos qos;
   const crane::grpc::QosInfo *qos_info = &request->qos();
 
@@ -551,6 +572,8 @@ grpc::Status CraneCtldServiceImpl::ModifyAccount(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   if (request->type() == crane::grpc::OperationType::Overwrite &&
       request->modify_field() ==
@@ -614,6 +637,8 @@ grpc::Status CraneCtldServiceImpl::ModifyUser(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   CraneExpected<void> modify_res;
 
@@ -754,6 +779,9 @@ grpc::Status CraneCtldServiceImpl::ModifyQos(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
+
   auto modify_res =
       g_account_manager->ModifyQos(request->uid(), request->name(),
                                    request->modify_field(), request->value());
@@ -775,6 +803,8 @@ grpc::Status CraneCtldServiceImpl::QueryAccountInfo(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   std::vector<Account> res_account_list;
   if (request->account_list().empty()) {
@@ -852,9 +882,8 @@ grpc::Status CraneCtldServiceImpl::QueryUserInfo(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
-
-  if (!CheckCertAndUIDAllowed_(context, request->uid()))
-    return grpc::Status{grpc::StatusCode::UNAUTHENTICATED, "cert err"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   std::unordered_set<std::string> user_list{request->user_list().begin(),
                                             request->user_list().end()};
@@ -935,6 +964,8 @@ grpc::Status CraneCtldServiceImpl::QueryQosInfo(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   std::vector<Qos> res_qos_list;
 
@@ -989,6 +1020,8 @@ grpc::Status CraneCtldServiceImpl::DeleteAccount(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   for (const auto &account_name : request->account_list()) {
     auto res = g_account_manager->DeleteAccount(request->uid(), account_name);
@@ -1013,6 +1046,9 @@ grpc::Status CraneCtldServiceImpl::DeleteUser(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
+
   for (const auto &user_name : request->user_list()) {
     auto res = g_account_manager->DeleteUser(request->uid(), user_name,
                                              request->account());
@@ -1038,6 +1074,8 @@ grpc::Status CraneCtldServiceImpl::DeleteQos(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   for (const auto &qos_name : request->qos_list()) {
     auto res = g_account_manager->DeleteQos(request->uid(), qos_name);
@@ -1064,6 +1102,9 @@ grpc::Status CraneCtldServiceImpl::BlockAccountOrUser(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status(grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready");
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
+
   CraneExpected<void> res;
   std::unordered_set<std::string> entity_list{request->entity_list().begin(),
                                               request->entity_list().end()};
@@ -1146,6 +1187,12 @@ grpc::Status CraneCtldServiceImpl::ResetUserCredential(
     grpc::ServerContext *context,
     const crane::grpc::ResetUserCredentialRequest *request,
     crane::grpc::ResetUserCredentialReply *response) {
+  if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
+    return grpc::Status{grpc::StatusCode::UNAVAILABLE,
+                        "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
+
   return Service::ResetUserCredential(context, request, response);
 }
 
@@ -1168,6 +1215,8 @@ grpc::Status CraneCtldServiceImpl::CreateReservation(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   auto res = g_account_manager->CheckUidIsAdmin(request->uid());
   if (!res) {
@@ -1187,6 +1236,8 @@ grpc::Status CraneCtldServiceImpl::DeleteReservation(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   auto res = g_account_manager->CheckUidIsAdmin(request->uid());
   if (!res) {
@@ -1253,6 +1304,8 @@ grpc::Status CraneCtldServiceImpl::EnableAutoPowerControl(
   if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
     return grpc::Status{grpc::StatusCode::UNAVAILABLE,
                         "CraneCtld Server is not ready"};
+  if (auto msg = CheckCertAndUIDAllowed_(context, request->uid()); msg)
+    return {grpc::StatusCode::UNAUTHENTICATED, msg.value()};
 
   CRANE_INFO(
       "Received enable auto power control request for {} nodes, enable: {}",
@@ -1310,32 +1363,38 @@ grpc::Status CraneCtldServiceImpl::EnableAutoPowerControl(
   return grpc::Status::OK;
 }
 
-bool CraneCtldServiceImpl::CheckCertAndUIDAllowed_(
+std::optional<std::string> CraneCtldServiceImpl::CheckCertAndUIDAllowed_(
     const grpc::ServerContext *context, uint32_t uid) {
-  if (!g_config.ListenConf.UseTls) return true;
+  if (!g_config.ListenConf.UseTls) return std::nullopt;
 
   auto cert = context->auth_context()->FindPropertyValues("x509_pem_cert");
-  if (cert.empty()) return false;
+  if (cert.empty()) return "Certificate is empty";
 
   std::string certificate = std::string(cert[0].data(), cert[0].size());
 
   auto result = util::ParseCertificate(certificate);
-  if (!result) return false;
+  if (!result) return "Certificate is invalid";
 
   if (!g_vault_client->IsCertAllowed(result.value().second))
-    return false;
+    return "Certificate has expired";
 
   std::vector<std::string> cn_parts = absl::StrSplit(result.value().first, '.');
   if (cn_parts.empty() || cn_parts[0].empty())
-    return false;
+    return "Certificate is invalid";
 
-  return static_cast<uint32_t>(std::stoul(cn_parts[0])) == uid;
+  if (static_cast<uint32_t>(std::stoul(cn_parts[0])) != uid)
+    return "Uid mismatch";
+
+  return std::nullopt;
 }
 
 grpc::Status CraneCtldServiceImpl::SignUserCertificate(
     grpc::ServerContext *context,
     const crane::grpc::SignUserCertificateRequest *request,
     crane::grpc::SignUserCertificateResponse *response) {
+  if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
+    return grpc::Status{grpc::StatusCode::UNAVAILABLE,
+                        "CraneCtld Server is not ready"};
   if (!g_config.ListenConf.tls_config.AllowedNodes.empty()) {
     std::string client_address = context->peer();
     std::vector<std::string> str_list = absl::StrSplit(client_address, ":");
